pyproject.toml
[tool.poetry]
name = "kuro"
version = "0.1.0"
description = "Kuro - Personal AI Assistant"
authors = ["kuro"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.scripts]
kuro = "src.main:main"

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
litellm = "^1.81.8"
rich = "^14.3.2"
prompt-toolkit = "^3.0.52"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
pyyaml = "^6.0.3"
aiosqlite = "^0.22.1"
aiofiles = "^25.1.0"
structlog = "^25.5.0"
keyring = "^25.7.0"
python-dotenv = "^1.2.1"
chromadb = "^1.4.1"
mss = "^10.1.0"
pillow = "^12.1.0"
playwright = "^1.58.0"
icalendar = "^6.3.2"
python-telegram-bot = {extras = ["ext"], version = "^22.6"}
fastapi = "^0.128.5"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
cryptography = ">=42.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
